ID指定での音声データ削除処理実装

+++ VoiceRecognition/src/main/java/com/example/demo/Controller/VoiceController.java
@@ -62,8 +62,10 @@ public class VoiceController {
     // --------------------------------------------------------------------
     @PostMapping(path = "select_delete")
     String select_delete(Model mov, @RequestParam("voiceID") String voiceID) {
-		int id = Integer.parseInt(voiceID);
-        service.Delete(id);
+        int id = Integer.parseInt(voiceID);
+        if( service.FindOne(id) != null){
+            service.Delete(id);
+        }
         return "redirect:/voice";
     }
     // --------------------------------------------------------------------
